services/recommendation-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.1
numpy==2.2.1
scikit-learn==1.6.1
pandas==2.2.3
joblib==1.4.2
redis==5.2.1
pymongo==4.10.1
psycopg2-binary==2.9.10
